requirements.txt
Django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
openpyxl==3.1.2
nepali-municipality==0.1.4